Feat response post

+++ src/main/java/site/sac/service/PostServiceImpl.java
@@ -4,11 +4,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.RequestBody;
 import site.sac.dto.PostDTO;
 import site.sac.mapper.PostMapper;
+import site.sac.mapper.UserLikeBoardMapper;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 @Slf4j
@@ -16,20 +18,25 @@ public class PostServiceImpl implements PostService{
     @Autowired
     private PostMapper postMapper;
 
+    @Autowired
+    private UserLikeBoardMapper userLikeBoardMapper;
+
     @Transactional
     @Override
     public void register(PostDTO postDTO) {
-        log.info("Post : " + postDTO.toString());
         postMapper.insert(postDTO);
     }
 
     @Override
-    public List<PostDTO> getAllPost() {
-        List<PostDTO> allPost = postMapper.getAllPost();
-        allPost.forEach(post->log.info("post : " + post.toString()));
+    public Map<String,Object>  getAllPost() {
 
-        return allPost;
+        List<PostDTO> allPost = postMapper.getAllPost();
+        Map<String,Object> result = new HashMap<>();
+        result.put("posts", allPost);
+        return result;
+
     }
+
 
     @Override
     public PostDTO getPostDetail(Long postId) {
@@ -38,11 +45,29 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public List<PostDTO> getPostsByBoardId(Long boardId) {
+    public Map<String,Object> getPostsByBoardId(Long boardId){
         List<PostDTO> postsByBoardId = postMapper.getAllPostsByBoardId(boardId);
-        postsByBoardId.forEach(post->log.info("post : " + post.toString()));
+        if (postsByBoardId == null){
+            throw new NullPointerException();
+        }
+        Map<String,Object> result = new HashMap<>();
+        result.put("posts", postsByBoardId);
+        result.put("count", postsByBoardId.size());
 
-        return postsByBoardId;
+        return result;
+    }
+    @Override
+    public Map<String,Object> getAllPostByUserId(Long userId){
+        List<String> userLikeBoard = userLikeBoardMapper.getAllByUserId(userId);
+        if (userLikeBoard == null){
+            throw new NullPointerException();
+        }
+        List<PostDTO> posts = postMapper.getAllPostByUserLikeBoard(userLikeBoard);
+        Map<String,Object> result = new HashMap<>();
+        result.put("posts", posts);
+        result.put("count", posts.size());
+        return result;
+
     }
 
     @Override
@@ -53,14 +78,20 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public PostDTO postEdit(PostDTO postDTO) {
-        postMapper.update(postDTO);
-        return postDTO;
+    public void postEdit(PostDTO postDTO) {
+        if(postMapper.read(postDTO.getPost_id()).getUser_id()== postDTO.getUser_id()){
+            postMapper.update(postDTO);
+
+        } else throw new NullPointerException();
+
     }
 
     @Override
-    public void delete(Long postId) {
-        postMapper.delete(postId);
+    public void delete(PostDTO postDTO, long postId) {
+        if(postMapper.read(postId).getUser_id()== postDTO.getUser_id()){
+            postMapper.delete(postId);
+
+        } else throw new NullPointerException();
     }
 
     @Override
